JAMES-2340 Rename contract tests so they can be played by JUnit
